fix: FriendList component
